requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-experimental==0.0.64
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.104
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.5
nodeenv==1.9.1
numpy==1.26.4
openai==1.42.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdoc3==0.11.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
protobuf==5.27.3
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
streamlit==1.37.1
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
yarl==1.9.4
zipp==3.20.1
